requirements.txt
absl-py==2.1.0
ale-py==0.8.1
array2gif==1.0.4
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
black==24.4.2
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.62.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.0.2
importlib_resources==6.3.0
intel-openmp==2021.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
lz4==4.3.3
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
minigrid==2.3.1
mkl==2021.4.0
moviepy==1.0.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.2
proglog==0.1.10
protobuf==4.25.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.12.0
pygame==2.5.2
pyglet==1.5.0
pyparsing==3.1.2
pyspark==3.5.0
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.10.1
Shimmy==0.2.1
six==1.16.0
sympy==1.12
tbb==2021.11.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.18.0
